upnp: Parameterize port mapping description
Applications may want to set their own mapping description, so add an
optional parameter to the openPort() method.

Similarly, when checking for a mapping using isMapped(), allow for an
optional description parameter that, if specified, is checked against
the description of an existing specific mapping. This can be used by
applications to check if they own an existing port mapping.